% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estBaseHazard.r
\name{estBaseHazard}
\alias{estBaseHazard}
\title{Estimate the baseline hazard function in the Cox model as a step function.}
\usage{
estBaseHazard(surv.data, Sllike, SurvParValue, Bi0, status, nblock)
}
\description{
Estimate the baseline hazard function in the Cox model as a step function.
}
