% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDisp.r
\name{estDisp}
\alias{estDisp}
\title{Estimate the dispersion parameters in the joint models, by maximizing the 
adjusted profile h-likelihood function.}
\usage{
estDisp(
  RespLog = list(Jlik1, Jlik2),
  Jdisp,
  sigma0,
  invSIGMA0,
  Dpars = c(Jraneff, Jfixed, Spar),
  long.data,
  surv.data,
  B,
  Bi,
  ParVal,
  Silent = T,
  Lval0 = NULL,
  lower,
  upper
)
}
\description{
Estimate the dispersion parameters in the joint models, by maximizing the 
adjusted profile h-likelihood function.
}
