% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estFixeff.r
\name{estFixeff}
\alias{estFixeff}
\title{Estimate the fixed parameters in the joint models, by maximizing the profile h-likelihood function.}
\usage{
estFixeff(
  RespLog = list(Jlik1, Jlik2),
  fixed,
  str_val,
  Dpars = Jraneff,
  long.data,
  surv.data,
  B,
  Bi,
  invSIGMA,
  sigma,
  ParVal = NULL,
  lower = NULL,
  Silent = T
)
}
\description{
Estimate the fixed parameters in the joint models, by maximizing the profile h-likelihood function.
}
