% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raneff.r
\name{get_raneff}
\alias{get_raneff}
\title{Estimate random effects}
\usage{
get_raneff(
  RespLog,
  long.data,
  surv.data,
  idVar,
  uniqueID,
  Jraneff,
  invSIGMA0,
  sigma0,
  n,
  ni,
  q,
  N,
  fixedest0,
  Silent,
  scale = T
)
}
\description{
Estimate random effects
}
