% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sd.r
\name{get_sd}
\alias{get_sd}
\title{Estimate the standard errors of fixed parameters}
\usage{
get_sd(
  long.data,
  surv.data,
  Bi,
  B,
  fixedest0,
  sigma0,
  invSIGMA0,
  RespLog,
  Jfixed,
  Jraneff,
  p,
  q,
  n
)
}
\description{
Estimate the standard errors of fixed parameters
}
