% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMfit.r
\name{JMfit}
\alias{JMfit}
\title{Fit a joint model}
\usage{
JMfit(
  glmeObject,
  survObject,
  long.data,
  surv.data,
  idVar,
  eventTime,
  survFit,
  method = c("h-likelihood", "aGH"),
  itertol = 0.001,
  Ptol = 0.01,
  epsilon = 1e-06,
  iterMax = 10,
  ghsize = 4,
  srcpath = NULL,
  parallel = F,
  Silent = T
)
}
\arguments{
\item{glmeObject}{A list, specifying the GLME models to be fitted.}

\item{survObject}{A list, specifying the survival model (either Cox PH or Weibull model) to be fitted.}

\item{long.data}{longitudinal data}

\item{surv.data}{survival data}

\item{idVar}{subject id}

\item{eventTime}{observed event time}

\item{survFit}{an object returned by coxph() or survreg() function to represent a fitted survival model from the two-step method}

\item{method}{a vector indicating which method to apply. If method='h-likelihood' (by default), the h-likelihood method is used; if method='aGH', the adaptive Gauss-Hermite method is used.}

\item{itertol}{Convergence tolerance on the relative absolute change in log-likelihood function between successive iterations. Convergence is declared when the change is less than itertol. Default is itertol = 0.001.}

\item{Ptol}{Convergence tolerance on the average relative absolute change in parameter estimates between successive iterations. Convergence is declared when the change is less than Ptol. Default is Ptol = 0.01.}

\item{epsilon}{A small numerical value, used to calculate the numerical value of the derivative of score function. The default value is 1e-6.}

\item{iterMax}{The maximum number of iterations. The default value is 10.}

\item{ghsize}{The number of quadrature points used in the adaptive GH method. The default value is 4.}

\item{Silent}{logical: indicating if messages about convergence success or failure should be suppressed.}
}
\description{
Fit a joint model
}
