% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMfit_HL.r
\name{JMfit_HL}
\alias{JMfit_HL}
\title{This function is for joint modelling of multiple (generalized) linear mixed-effect 
models and a survival model (Cox PH or Weibull), via h-likelihood method. It also 
addresses left-truncation of a continuous longitudinal data fitted by a LME model.}
\usage{
JMfit_HL(
  glmeObject,
  survObject,
  long.data,
  surv.data,
  idVar,
  eventTime,
  survFit,
  itertol = 0.001,
  Ptol = 0.01,
  iterMax = 10,
  Silent = T
)
}
\description{
This function is for joint modelling of multiple (generalized) linear mixed-effect 
models and a survival model (Cox PH or Weibull), via h-likelihood method. It also 
addresses left-truncation of a continuous longitudinal data fitted by a LME model.
}
